+++ src/main/java/com/tienda/inventario/repositories/ProductoRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.repository.query.Param;
 
 import com.tienda.inventario.entities.Producto;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {
 
     // Productos con stock > 0
-    List<Producto> findByCantidadGreaterThan(Long cantidad);
+    List<Producto> findByCantidadGreaterThan(Integer cantidad);
 
     // Productos por rango de fecha de última compra
     List<Producto> findByUltimaCompraBetween(LocalDateTime desde, LocalDateTime hasta);
@@ -21,7 +21,7 @@ public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
     // Precio del producto por id
 @Query("select p.precio from Producto p where p.id = :id")
-Float findPrecioById(@Param("id") Long id);
+Float findPrecioById(@Param("id") Integer id);
 
 
 }
